# backend/docs/swagger.yaml
openapi: 3.0.0
info:
  title: GPASS API
  version: 1.0.0
  description: |-
    This is the API documentation for the GPASS project.
  contact:
    email: hahaisme@haha.WhySoSerious
...
paths:
...
  /api/1.0/exams/details/{examId}:
    get:
      tags:
        - exams
      summary: Get exam details
      description: Use this request to get exam details
      operationId: getExamDetails
      parameters:
        - name: examId
          in: path
          description: The exam ID
          required: true
          schema:
            type: string
          example: 14
      responses:
        "200":
          description: Exam details retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                type: string
                description: The error message
              example:
                message: Bad request
        "404":
          description: Exam not found
          content:
            application/json:
              schema:
                type: string
                description: The error message
              example:
                message: Exam not found
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: string
                description: The error message
              example:
                message: Internal server error
    put:
      tags:
        - exams
      summary: Update exam details
      description: Use this request to update exam details
      operationId: updateExamDetails
      parameters:
        - name: examId
          in: path
          description: The exam ID
          required: true
          schema:
            type: string
          example: 14
      requestBody:
        description: Exam object that needs to be updated
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                classId:
                  type: string
                  description: The class ID of the exam
                  example: 9
                className:
                  type: string
                  description: The class name of the exam
                type:
                  type: string
                  description: The type of the exam
                year:
                  type: string
                  description: The year of the exam
                teacher:
                  type: string
                  description: The teacher of the exam
                main_file:
                  type: string
                  format: binary
                  description: The main image of the exam
                ans_file:
                  type: string
                  format: binary
                  description: The images of the exam
                sheet_files:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: The images of the exam
                has_ans:
                  type: boolean
                  description: Whether the exam has answer
                  example: true
              required:
                - classId
              #   - type
              #   - year
              #   - teacher
              #   - main_file
              #   - has_ans
      responses:
        "200":
          description: Exam details updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                type: string
                description: The error message
              example:
                message: Bad request
        "404":
          description: Exam not found
          content:
            application/json:
              schema:
                type: string
                description: The error message
              example:
                message: Exam not found
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: string
                description: The error message
              example:
                message: Internal server error
    delete:
      tags:
        - exams
      summary: Delete exam
      description: Use this request to delete exam
      operationId: deleteExam
      parameters:
        - name: examId
          in: path
          description: The exam ID
          required: true
          schema:
            type: string
          example: 14
      responses:
        "200":
          description: Exam deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                type: string
                description: The error message
              example:
                message: Bad request
        "404":
          description: Exam not found
          content:
            application/json:
              schema:
                type: string
                description: The error message
              example:
                message: Exam not found
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: string
                description: The error message
              example:
                message: Internal server error
...
components:
  schemas:
    Exam:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
          description: The exam ID
        type:
          type: string
          example: midterm
          description: The type of the exam
        teacher:
          type: string
          example: Haha
          description: The teacher of the exam
        year:
          type: string
          example: 2020
          description: The year of the exam
        rating_id:
          type: integer
          format: int64
          example: 1
          description: The rating ID of the exam
        main_file:
          type: string
          example: https://example.com/file.pdf
          description: The main file of the exam
        ans_file:
          type: string
          example: https://example.com/file.pdf
          description: The answer file of the exam
        sheet_files:
          type: array
          items:
            type: string
            example: https://example.com/file.pdf
          description: The sheet files of the exam
        has_ans:
          type: boolean
          example: true
          description: Whether the exam has answer
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT